Added cd command and project utilities

+++ tracker/utils/print_lists.py
@@ -14,8 +14,8 @@ def print_two_coloumn(l):
     if len(l) % 2 != 0:
         l.append(" ")
 
-    split = len(l) / 2
-    l1 = l[0: split]
+    split = int(len(l) / 2)
+    l1 = l[0:split]
     l2 = l[split:]
     for key, value in zip(l1, l2):
         print("{0:<20s} {1}".format(key, value))
